Refactored API signatures to use a tokenizer.
I'm looking into introducing a feature where the API Dump and the API
Diff Logs can be presented as an HTML file with CSS controlling the
syntax highlighting. In order to preserve consistency between the two
formats, I have created a system that allows each descriptor to map out
a schema for generating their signature. Values for each token in the
schema are provided per-instance.